Time: 29 ms (65.31%), Space: 10.8 MB (74.27%) - LeetHub

+++ 115-distinct-subsequences/115-distinct-subsequences.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int numDistinct(string s, string t) {
-        unsigned long dp [s.length()+1][t.length()+1];
+        unsigned int dp [s.length()+1][t.length()+1];
         int m=s.length();
         int n=t.length();
         int i,j;
